AbstractTests - Add more examples

+++ src/main/java/library/Page.java
@@ -2,8 +2,7 @@ package library;
 
 public class Page {
 
-	public Page(Long pageNumber) {
-
+	public Page(int pageNumber) {
 	}
 
 }
